Add YorkieCountable.swift
