+++ Sources/Document/Json/JSONCounter.swift
@@ -16,14 +16,6 @@
 
 import Foundation
 
-public protocol YorkieCountable: BinaryInteger {
-    var littleEndian: Self { get }
-    func addingReportingOverflow(_ other: Self) -> (partialValue: Self, overflow: Bool)
-}
-
-extension Int32: YorkieCountable {}
-extension Int64: YorkieCountable {}
-
 /**
  * `JSONCounter` is a custom data type that is used to counter.
  */
